% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landuse.R
\docType{data}
\name{land_use_lookup}
\alias{land_use_lookup}
\title{Landuse lookup codes}
\format{
An object of class \code{data.frame} with 193 rows and 8 columns.
}
\usage{
land_use_lookup
}
\description{
Landuse lookup codes
}
\references{
\url{https://data.gov.au/dataset/ds-dga-8d5d0a09-d100-407b-b326-6e775025feee/details?q=land\%20use\%20catchment}
}
\keyword{data}
