% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landuse.R
\name{landuse_wms}
\alias{landuse_wms}
\title{Get Landuse Tiles}
\usage{
landuse_wms(catchment_sf)
}
\arguments{
\item{catchment_sf}{sf polygon of catchments to have landuse extracted for them}
}
\value{
raster
}
\description{
Get Landuse Tiles
}
