% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackcatcment.R
\name{track_catchments}
\alias{track_catchments}
\title{Track catchments upstream}
\usage{
track_catchments(point)
}
\arguments{
\item{point}{sf point of location to begin at}
}
\value{
sf/data.frame
}
\description{
Track catchments upstream
}
